progress counter now uses size_t

+++ src/global/progresscounter.cpp
@@ -12,16 +12,16 @@ ProgressCounter::ProgressCounter(QObject *parent)
     : QObject(parent)
 {}
 
-void ProgressCounter::increaseTotalStepsBy(uint32_t steps)
+void ProgressCounter::increaseTotalStepsBy(size_t steps)
 {
     m_totalSteps += steps;
     step(0u);
 }
 
-void ProgressCounter::step(const uint32_t steps)
+void ProgressCounter::step(const size_t steps)
 {
     m_steps += steps;
-    const uint32_t percentage = (m_totalSteps == 0u) ? 0u : (100u * m_steps / m_totalSteps);
+    const size_t percentage = (m_totalSteps == 0u) ? 0u : (100u * m_steps / m_totalSteps);
     if (percentage != m_percentage) {
         m_percentage = percentage;
         emit sig_onPercentageChanged(percentage);
